+++ HISTORY.rst
@@ -16,6 +16,7 @@ Release History
 **BugFixes**
 
 - Fix the usability of the `fmt` argument in `jupytext.read` (#312)
+- Fix the download notebook error when `c.notebook_extensions` has a custom value (#318)
 
 1.2.1 (2019-07-20)
 ++++++++++++++++++++++